update list emails newsletter fix for colloque liste and add emails to list

+++ app/Droit/Newsletter/Repo/NewsletterListEloquent.php
@@ -81,9 +81,11 @@ class NewsletterListEloquent implements NewsletterListInterface{
         $list->fill($data);
         $list->save();
 
-        if(isset($data['emails']) && !empty($data['emails']))
-        {
-            $list->emails()->delete();
+        if(isset($data['emails']) && !empty($data['emails'])) {
+
+            if(isset($data['update'])){
+                $list->emails()->delete();
+            }
 
             foreach($data['emails'] as $email) {
                 $list->emails()->save(new \App\Droit\Newsletter\Entities\Newsletter_emails(['list_id' => $list->id, 'email' => $email]));
